# Conflicts:
#	build.gradle
#	src/main/java/f3f/domain/comment/domain/Comment.java
#	src/main/java/f3f/domain/likes/domain/Likes.java
#	src/main/java/f3f/domain/scrap/domain/Scrap.java

+++ src/main/java/f3f/domain/board/application/BoardService.java
@@ -4,14 +4,23 @@ import f3f.domain.board.dao.BoardRepository;
 import f3f.domain.board.domain.Board;
 import f3f.domain.board.dto.BoardDTO;
 import f3f.domain.board.dto.BoardDTO.BoardInfoDTO;
-import f3f.domain.board.exception.NotFoundBoardException;
+import f3f.domain.board.exception.BoardMissMatchUserException;
 import f3f.domain.board.exception.NotFoundBoardCategoryException;
+import f3f.domain.board.exception.NotFoundBoardException;
 import f3f.domain.board.exception.NotFoundBoardUserException;
+import f3f.domain.category.dao.CategoryRepository;
+import f3f.domain.category.domain.Category;
+import f3f.domain.category.exception.NotFoundCategoryException;
+import f3f.domain.user.dao.MemberRepository;
+import f3f.domain.user.domain.Member;
+import f3f.domain.user.exception.MemberNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
+
 /*
  * 필요한 기능
  * 1. 게시글 저장
@@ -28,7 +37,8 @@ import java.util.List;
 public class BoardService {
 
     private final BoardRepository boardRepository;
-
+    private final CategoryRepository categoryRepository;
+    private final MemberRepository memberRepository;
     @Transactional
     public void saveBoard(BoardDTO.SaveRequest request) {
 
@@ -36,7 +46,7 @@ public class BoardService {
             throw new NotFoundBoardCategoryException("게시글에 카테고리 정보가 존재하지 않습니다.");
         }
 
-        if (request.getUser() == null){
+        if (request.getMember() == null){
             throw new NotFoundBoardUserException("게시글에 유저 정보가 존재하지 않습니다.");
         }
 
@@ -44,9 +54,16 @@ public class BoardService {
     }
 
     @Transactional
-    public Board updateBoard(long boardId, BoardDTO.SaveRequest request) {
+    public Board updateBoard(long boardId, long userId, BoardDTO.SaveRequest request) {
+        Member member = memberRepository.findById(userId)
+                .orElseThrow(MemberNotFoundException::new);
+
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(() -> new NotFoundBoardException("존재하지 않는 게시글입니다."));
+
+        if (member.getId() != board.getMember().getId()){
+            throw new BoardMissMatchUserException();
+        }
 
         board.updateBoard(request);
 
@@ -54,22 +71,32 @@ public class BoardService {
     }
 
     @Transactional
-    public Board deleteBoard(long boardId) {
+    public Board deleteBoard(long boardId, long userId) {
+        Member member = memberRepository.findById(userId)
+                .orElseThrow(MemberNotFoundException::new);
 
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(() -> new NotFoundBoardException("존재하지 않는 게시글입니다."));
 
+        if (member.getId() != board.getMember().getId()){
+            throw new BoardMissMatchUserException();
+        }
         boardRepository.deleteById(board.getId());
 
         return board;
     }
 
     @Transactional(readOnly = true)
-    public BoardInfoDTO getBoardInfo(long boardId) {
+    public BoardInfoDTO getBoardInfo(long boardId , long userId) {
+        Member member = memberRepository.findById(userId)
+                .orElseThrow(MemberNotFoundException::new);
 
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(NotFoundBoardException::new);
 
+        if (member.getId() != board.getMember().getId()){
+            throw new BoardMissMatchUserException();
+        }
         return board.toBoardInfoDTO();
     }
 
@@ -86,6 +113,12 @@ public class BoardService {
      */
     @Transactional(readOnly = true)
     public List<BoardInfoDTO> getBoardListByCategoryId(long categoryId){
-        return null;
+        Category category = categoryRepository.findById(categoryId)
+                .orElseThrow(NotFoundCategoryException::new);
+
+        List<BoardInfoDTO> boardInfoList = category.getBoardList().stream()
+                .map(Board::toBoardInfoDTO).collect(Collectors.toList());
+
+        return boardInfoList;
     }
 }
